+++ src/redux/action-creators.js
@@ -1435,3 +1435,26 @@ export function deleteUndoEntry(id) {
     payload: { id },
   };
 }
+
+export function pinPost(postId, target) {
+  return {
+    type: ActionTypes.PIN_POST,
+    apiRequest: Api.pinPost,
+    payload: { postId, target },
+  };
+}
+
+export function unpinPost(postId, target) {
+  return {
+    type: ActionTypes.UNPIN_POST,
+    apiRequest: Api.unpinPost,
+    payload: { postId, target },
+  };
+}
+
+export function reorderFeedEntries(entries) {
+  return {
+    type: ActionTypes.REORDER_FEED_ENTRIES,
+    payload: { entries },
+  };
+}