+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.140.0] - Not released
+### Added
+- Pinning posts feature. Users can now pin/unpin posts at the top of their own
+  feed or the feed of a group they administer.
 
 ## [1.139.2] - 2025-08-04
 ### Added